+++ bundles/core/org.eclipse.smarthome.core/ESH-INF/config/locale.xml
@@ -4,7 +4,7 @@
 	xsi:schemaLocation="http://eclipse.org/smarthome/schemas/config-description/v1.0.0
         http://eclipse.org/smarthome/schemas/config-description-1.0.0.xsd">
 
-	<config-description uri="system:locale">
+	<config-description uri="system:i18n">
 		<parameter name="language" type="text">
 			<label>Language</label>
 			<description><![CDATA[
