--- bundles/core/org.eclipse.smarthome.core/OSGI-INF/localeprovider.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (c) 2014-2017 by the respective copyright holders.
-    All rights reserved. This program and the accompanying materials
-    are made available under the terms of the Eclipse Public License v1.0
-    which accompanies this distribution, and is available at
-    http://www.eclipse.org/legal/epl-v10.html
-
--->
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" enabled="true" immediate="true" modified="modified" name="org.eclipse.smarthome.core.localeprovider">
-   <implementation class="org.eclipse.smarthome.core.internal.i18n.LocaleProviderImpl"/>
-   <service>
-      <provide interface="org.eclipse.smarthome.core.i18n.LocaleProvider"/>
-   </service>
-   <property name="service.config.description.uri" type="String" value="system:locale"/>
-   <property name="service.config.label" type="String" value="Locale"/>
-   <property name="service.config.category" type="String" value="system"/>
-</scr:component>
