+++ bundles/core/org.eclipse.smarthome.core/src/main/java/org/eclipse/smarthome/core/internal/I18nProviderImpl.java
@@ -1,0 +1,162 @@
+package org.eclipse.smarthome.core.internal;
+
+import java.text.MessageFormat;
+import java.util.Locale;
+import java.util.Map;
+import java.util.ResourceBundle;
+
+import org.apache.commons.lang.StringUtils;
+import org.eclipse.smarthome.core.i18n.LocaleProvider;
+import org.eclipse.smarthome.core.i18n.LocationProvider;
+import org.eclipse.smarthome.core.i18n.TranslationProvider;
+import org.eclipse.smarthome.core.internal.i18n.LanguageResourceBundleManager;
+import org.eclipse.smarthome.core.internal.i18n.ResourceBundleTracker;
+import org.eclipse.smarthome.core.library.types.PointType;
+import org.osgi.framework.Bundle;
+import org.osgi.service.component.ComponentContext;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+/**
+ * The {@link I18nProviderImpl} is a concrete implementation of the {@link TranslationProvider}, {@link LocaleProvider},
+ * and {@link LocationProvider} service interfaces.
+ * *
+ * <p>
+ * This implementation uses the i18n mechanism of Java ({@link ResourceBundle}) to translate a given key into text. The
+ * resources must be placed under the specific directory {@link LanguageResourceBundleManager#RESOURCE_DIRECTORY} within
+ * the certain modules. Each module is tracked in the platform by using the {@link ResourceBundleTracker} and managed by
+ * using one certain {@link LanguageResourceBundleManager} which is responsible for the translation.
+ * <p>
+ * <p>
+ * It reads a user defined configuration to set a locale and a location for this installation
+ *
+ * @author Michael Grammling - Initial Contribution of TranslationProvider
+ * @author Thomas Höfer - Added getText operation with arguments
+ * @author Markus Rathgeb - Initial contribution and API of LocaleProvider
+ * @author Stefan Triller - Initial contribution and API of LocationProvider
+ *
+ */
+public class I18nProviderImpl implements TranslationProvider, LocaleProvider, LocationProvider {
+
+    private final Logger logger = LoggerFactory.getLogger(getClass());
+
+    // LocaleProvider
+    private static final String LANGUAGE = "language";
+    private static final String SCRIPT = "script";
+    private static final String REGION = "region";
+    private static final String VARIANT = "variant";
+    private Locale locale;
+
+    // TranslationProvider
+    private ResourceBundleTracker resourceBundleTracker;
+
+    // LocationProvider
+    private static final String LOCATION = "location";
+    private PointType location;
+
+    @SuppressWarnings("unchecked")
+    protected void activate(ComponentContext componentContext) {
+        modified((Map<String, Object>) componentContext.getProperties());
+
+        this.resourceBundleTracker = new ResourceBundleTracker(componentContext.getBundleContext(), this);
+        this.resourceBundleTracker.open();
+    }
+
+    protected void deactivate(ComponentContext componentContext) {
+        this.resourceBundleTracker.close();
+    }
+
+    protected synchronized void modified(Map<String, Object> config) {
+        final String language = (String) config.get(LANGUAGE);
+        final String script = (String) config.get(SCRIPT);
+        final String region = (String) config.get(REGION);
+        final String variant = (String) config.get(VARIANT);
+        final String location = (String) config.get(LOCATION);
+
+        if (location != null) {
+            try {
+                this.location = PointType.valueOf(location);
+            } catch (IllegalArgumentException e) {
+                // preserve old location or null if none was set before
+                logger.warn("Could not set new location, keeping old one: ", e.getMessage());
+            }
+        }
+
+        if (StringUtils.isEmpty(language)) {
+            // at least the language must be defined otherwise the system default locale is used
+            logger.debug("No language set, fallback to default system locale");
+            locale = null;
+            return;
+        }
+
+        final Locale.Builder builder = new Locale.Builder();
+        try {
+            builder.setLanguage(language);
+        } catch (final RuntimeException ex) {
+            logger.warn("Language ({}) is invalid. Cannot create locale, keep old one.", language, ex);
+            return;
+        }
+
+        try {
+            builder.setScript(script);
+        } catch (final RuntimeException ex) {
+            logger.warn("Script ({}) is invalid. Skip it.", script, ex);
+            return;
+        }
+
+        try {
+            builder.setRegion(region);
+        } catch (final RuntimeException ex) {
+            logger.warn("Region ({}) is invalid. Skip it.", region, ex);
+            return;
+        }
+
+        try {
+            builder.setVariant(variant);
+        } catch (final RuntimeException ex) {
+            logger.warn("Variant ({}) is invalid. Skip it.", variant, ex);
+            return;
+        }
+
+        locale = builder.build();
+        logger.info("Locale set to {}", locale);
+    }
+
+    @Override
+    public PointType getLocation() {
+        return location;
+    }
+
+    @Override
+    public Locale getLocale() {
+        if (locale == null) {
+            return Locale.getDefault();
+        }
+        return locale;
+    }
+
+    @Override
+    public String getText(Bundle bundle, String key, String defaultText, Locale locale) {
+        LanguageResourceBundleManager languageResource = this.resourceBundleTracker.getLanguageResource(bundle);
+
+        if (languageResource != null) {
+            String text = languageResource.getText(key, locale);
+            if (text != null) {
+                return text;
+            }
+        }
+
+        return defaultText;
+    }
+
+    @Override
+    public String getText(Bundle bundle, String key, String defaultText, Locale locale, Object... arguments) {
+        String text = getText(bundle, key, defaultText, locale);
+
+        if (text != null) {
+            return MessageFormat.format(text, arguments);
+        }
+
+        return text;
+    }
+}
